added policies and simple spell check

+++ app/Domain/Notes/NoteController.php
@@ -3,12 +3,15 @@
 namespace App\Domain\Notes;
 
 use App\Domain\Notes\Repositories\NoteRepository;
+use App\Domain\Notes\Requests\SpellCheckRequest;
 use App\Domain\Notes\Requests\StoreNoteRequest;
 use App\Domain\Notes\Requests\UpdateNoteRequest;
 use App\Domain\Notes\Resources\HtmlNoteResource;
 use App\Domain\Notes\Resources\NoteResource;
 use App\Http\Controllers\Controller;
+use App\Services\LanguageToolService;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class NoteController extends Controller
 {
@@ -29,17 +32,26 @@ class NoteController extends Controller
 
     public function show(Note $note)
     {
+        if (Auth::user()->cannot('view', $note)) {
+            abort(403, 'You can not view this note.');
+        }
         return new NoteResource($note);
     }
 
     public function update(UpdateNoteRequest $request, Note $note)
     {
+        if (Auth::user()->cannot('update', $note)) {
+            abort(403, 'You can not update this note.');
+        }
         $noteData = $request->validated();
         return new NoteResource($this->repository->update($noteData, $note));
     }
 
     public function destroy(Note $note)
     {
+        if (Auth::user()->cannot('delete', $note)) {
+            abort(403, 'You can not delete this note.');
+        }
         $note->delete();
         return response()->json(
             ['message' => 'Note has been deleted']
@@ -48,6 +60,16 @@ class NoteController extends Controller
 
     public function render(Note $note)
     {
+        if (Auth::user()->cannot('delete', $note)) {
+            abort(403, 'You can not render this note.');
+        }
         return new HtmlNoteResource($note);
+    }
+
+    public function spellCheck(Note $note, SpellCheckRequest $request, LanguageToolService $service)
+    {
+        $language = $request->validated()['language'] ?? 'auto';
+        $response = $service->checkGrammar($note->content, $language);
+        return response()->json($response);
     }
 }
